+++ src/main/java/mpicbg/pointdescriptor/LinkedPoint.java
@@ -1,6 +1,6 @@
 package mpicbg.pointdescriptor;
 
-import mpicbg.imglib.util.Util;
+import net.imglib2.util.Util;
 import mpicbg.models.Point;
 
 public class LinkedPoint<P> extends Point
@@ -9,13 +9,13 @@ public class LinkedPoint<P> extends Point
 
 	final P link;
 	
-	public LinkedPoint( final float[] l, final P link )
+	public LinkedPoint( final double[] l, final P link )
 	{
 		super( l.clone() );		
 		this.link = link;
 	}
 
-	public LinkedPoint( final float[] l, final float[] w, final P link )
+	public LinkedPoint( final double[] l, final double[] w, final P link )
 	{
 		super( l.clone(), w.clone() );
 		this.link = link;
